numpy

+++ deepdrivewe/examples/synd_ntl9_lof/inference.py
@@ -150,7 +150,7 @@ def run_inference(  # noqa: PLR0915
     data = [sim.data for sim in input_data]
     contact_maps = []
     for sim_data in data:
-        contact_maps.extend(sim_data['contact_maps'])
+        contact_maps.extend(sim_data['contact_maps'].astype(np.int16))
     # contact_maps = np.array(cmaps, dtype=object)
 
     print(f'{len(contact_maps)=}')
